+++ src/free_glyph.c
@@ -7,14 +7,17 @@ void free_glyph_atlas_init(Free_Glyph_Atlas *atlas, FT_Face face)
     // TODO: Introduction of SDF font slowed down the start up time
     // We need to investigate what's up with that
     FT_Int32 load_flags = FT_LOAD_RENDER | FT_LOAD_TARGET_(FT_RENDER_MODE_SDF);
-    for (int i = 32; i < 128; ++i) {
-        if (FT_Load_Char(face, i, load_flags)) {
+    for (int i = 32; i < 128; ++i)
+    {
+        if (FT_Load_Char(face, i, load_flags))
+        {
             fprintf(stderr, "ERROR: could not load glyph of a character with code %d\n", i);
             exit(1);
         }
 
         atlas->atlas_width += face->glyph->bitmap.width;
-        if (atlas->atlas_height < face->glyph->bitmap.rows) {
+        if (atlas->atlas_height < face->glyph->bitmap.rows)
+        {
             atlas->atlas_height = face->glyph->bitmap.rows;
         }
     }
@@ -33,32 +36,35 @@ void free_glyph_atlas_init(Free_Glyph_Atlas *atlas, FT_Face face)
         GL_TEXTURE_2D,
         0,
         GL_RED,
-        (GLsizei) atlas->atlas_width,
-        (GLsizei) atlas->atlas_height,
+        (GLsizei)atlas->atlas_width,
+        (GLsizei)atlas->atlas_height,
         0,
         GL_RED,
         GL_UNSIGNED_BYTE,
         NULL);
 
     int x = 0;
-    for (int i = 32; i < 128; ++i) {
-        if (FT_Load_Char(face, i, load_flags)) {
+    for (int i = 32; i < 128; ++i)
+    {
+        if (FT_Load_Char(face, i, load_flags))
+        {
             fprintf(stderr, "ERROR: could not load glyph of a character with code %d\n", i);
             exit(1);
         }
 
-        if (FT_Render_Glyph(face->glyph, FT_RENDER_MODE_NORMAL)) {
+        if (FT_Render_Glyph(face->glyph, FT_RENDER_MODE_NORMAL))
+        {
             fprintf(stderr, "ERROR: could not render glyph of a character with code %d\n", i);
             exit(1);
         }
 
-        atlas->metrics[i].ax = face->glyph->advance.x >> 6;
-        atlas->metrics[i].ay = face->glyph->advance.y >> 6;
-        atlas->metrics[i].bw = face->glyph->bitmap.width;
-        atlas->metrics[i].bh = face->glyph->bitmap.rows;
-        atlas->metrics[i].bl = face->glyph->bitmap_left;
-        atlas->metrics[i].bt = face->glyph->bitmap_top;
-        atlas->metrics[i].tx = (float) x / (float) atlas->atlas_width;
+        atlas->metrics[i].ax = (float)(face->glyph->advance.x >> 6);
+        atlas->metrics[i].ay = (float)(face->glyph->advance.y >> 6);
+        atlas->metrics[i].bw = (float)face->glyph->bitmap.width;
+        atlas->metrics[i].bh = (float)face->glyph->bitmap.rows;
+        atlas->metrics[i].bl = (float)face->glyph->bitmap_left;
+        atlas->metrics[i].bt = (float)face->glyph->bitmap_top;
+        atlas->metrics[i].tx = (float)x / (float)atlas->atlas_width;
 
         glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
         glTexSubImage2D(
@@ -77,13 +83,16 @@ void free_glyph_atlas_init(Free_Glyph_Atlas *atlas, FT_Face face)
 
 float free_glyph_atlas_cursor_pos(const Free_Glyph_Atlas *atlas, const char *text, size_t text_size, Vec2f pos, size_t col)
 {
-    for (size_t i = 0; i < text_size; ++i) {
-        if (i == col) {
+    for (size_t i = 0; i < text_size; ++i)
+    {
+        if (i == col)
+        {
             return pos.x;
         }
 
         size_t glyph_index = text[i];
-        if (glyph_index >= GLYPH_METRICS_CAPACITY) {
+        if (glyph_index >= GLYPH_METRICS_CAPACITY)
+        {
             glyph_index = '?';
         }
 
@@ -97,10 +106,12 @@ float free_glyph_atlas_cursor_pos(const Free_Glyph_Atlas *atlas, const char *tex
 
 void free_glyph_atlas_measure_line_sized(Free_Glyph_Atlas *atlas, const char *text, size_t text_size, Vec2f *pos)
 {
-    for (size_t i = 0; i < text_size; ++i) {
+    for (size_t i = 0; i < text_size; ++i)
+    {
         size_t glyph_index = text[i];
         // TODO: support for glyphs outside of ASCII range
-        if (glyph_index >= GLYPH_METRICS_CAPACITY) {
+        if (glyph_index >= GLYPH_METRICS_CAPACITY)
+        {
             glyph_index = '?';
         }
         Glyph_Metric metric = atlas->metrics[glyph_index];
@@ -112,17 +123,19 @@ void free_glyph_atlas_measure_line_sized(Free_Glyph_Atlas *atlas, const char *te
 
 void free_glyph_atlas_render_line_sized(Free_Glyph_Atlas *atlas, Simple_Renderer *sr, const char *text, size_t text_size, Vec2f *pos, Vec4f color)
 {
-    for (size_t i = 0; i < text_size; ++i) {
+    for (size_t i = 0; i < text_size; ++i)
+    {
         size_t glyph_index = text[i];
         // TODO: support for glyphs outside of ASCII range
-        if (glyph_index >= GLYPH_METRICS_CAPACITY) {
+        if (glyph_index >= GLYPH_METRICS_CAPACITY)
+        {
             glyph_index = '?';
         }
         Glyph_Metric metric = atlas->metrics[glyph_index];
         float x2 = pos->x + metric.bl;
         float y2 = -pos->y - metric.bt;
-        float w  = metric.bw;
-        float h  = metric.bh;
+        float w = metric.bw;
+        float h = metric.bh;
 
         pos->x += metric.ax;
         pos->y += metric.ay;
@@ -132,7 +145,7 @@ void free_glyph_atlas_render_line_sized(Free_Glyph_Atlas *atlas, Simple_Renderer
             vec2f(x2, -y2),
             vec2f(w, -h),
             vec2f(metric.tx, 0.0f),
-            vec2f(metric.bw / (float) atlas->atlas_width, metric.bh / (float) atlas->atlas_height),
+            vec2f(metric.bw / (float)atlas->atlas_width, metric.bh / (float)atlas->atlas_height),
             color);
     }
 }
